added transactiontemplate, sql injection to class, added cleanup.sql file, edited application-test.yml
