+++ src/test/resources/cleanup.sql
@@ -1,0 +1,4 @@
+USE discount_web_app;
+SET FOREIGN_KEY_CHECKS = 0;
+TRUNCATE TABLE discount_web_app.role;
+SET FOREIGN_KEY_CHECKS = 1;
